portable_hello: Add example of using standard portable POSIX source.
Add an example hello world that shows how a "normal" POSIX program can
be used with RTEMS. The pattern of putting RTEMS initialisation and
configuration in a dedicated file (e.g. rtems_config.c) is done in
some other examples but not in a simple one just to highlight this pattern.